Tech_urban_growth/requirements.txt
requests==2.31.0
pymongo
streamlit
pandas
requests
pymongo
psycopg2-binary
apache-airflow-providers-postgres
apache-airflow-providers-sftp
apache-airflow-providers-ssh

